refactor(diagnostics): split duplicate hash key lint (#0000)
